又在一次修改了version函数，以及一些警告信息

+++ inc/nanai_ann_nannmgr.h
@@ -42,7 +42,7 @@ namespace nanai {
                            nanai_ann_nanncalc *calc);
     virtual void waits();
     virtual void set_max(int max);
-    virtual int version() const;
+    virtual double version() const;
     
   protected:
     virtual void configure();
